+++ index.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" /> 
-        <script defer type="module" src="./ui.js"></script>       
+        <script defer type="module" src="./index.js"></script>       
         <title>Test</title>
     </head>
     <body>
@@ -11,7 +11,8 @@
     <div class="questionBlock" style="width: 500px; margin: 0 auto;">
         <div class="questionBlock__question">
             <p class="questionBlock__text"></p>
-            <form class="questionBlock__form" onsubmit="handleSubmit"></form>            
+            <form class="questionBlock__form" onsubmit="handleSubmit"></form>   
+            <p class="questionBlock__index"></p>         
         
         <div class="questionBlock__buttonBlock">
             <button type="button" class="questionBlock__start">Start</button>
